доработана связь с БД с nvarchar(450) userid

+++ TEL_ProjectBus/DAL/Entities/Common/ClassifierKey.cs
@@ -6,6 +6,6 @@
 /// </summary>
 public readonly record struct ClassifierKey(int Value)
 {
-	public static implicit operator long(ClassifierKey id) => id.Value;
+	public static implicit operator int(ClassifierKey id) => id.Value;
 	public static explicit operator ClassifierKey(int value) => new(value);
 }
